+++ springgear-core/src/main/java/org/springgear/EnableSpringGear.java
@@ -2,9 +2,7 @@ package org.springgear;
 
 import org.springframework.context.annotation.Import;
 import org.springframework.core.annotation.AliasFor;
-import org.springgear.core.beans.AbstractSpringGearProxyProcessor;
-import org.springgear.core.beans.DefaultBeanDefinitionProcessor;
-import org.springgear.core.context.registrars.SpringGearBeanAnnotationRegistrar;
+import org.springgear.core.context.registrar.SpringGearBeanAnnotationRegistrar;
 import org.springgear.core.engine.execute.AbstractSpringGearEngineExecutor;
 import org.springgear.core.engine.execute.DefaultSpringGearEngineExecutor;
 
@@ -32,13 +30,6 @@ public @interface EnableSpringGear {
      */
     @AliasFor("value")
     String[] basePackages() default {};
-
-    /**
-     * 代理处理器
-     *
-     * @return
-     */
-    Class<? extends AbstractSpringGearProxyProcessor> processor() default DefaultBeanDefinitionProcessor.class;
 
     /**
      * 执行过程处理器
